Basic authentication

+++ src/Controller/ArtistController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[Route('/artist')]
+#[IsGranted('ROLE_ADMIN')]
 class ArtistController extends AbstractController
 {
     #[Route('/', name: 'app_artist_index', methods: ['GET'])]
@@ -54,6 +55,7 @@ class ArtistController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_artist_edit', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_SUPER_ADMIN', message: 'You are not allowed to access the admin dashboard.')]
     public function edit(Request $request, Artist $artist, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(ArtistType::class, $artist);
@@ -72,8 +74,11 @@ class ArtistController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_artist_delete', methods: ['POST'])]
+    #[IsGranted('ROLE_SUPER_ADMIN', message: 'You are not allowed to access the admin dashboard.')]
     public function delete(Request $request, Artist $artist, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        
         if ($this->isCsrfTokenValid('delete'.$artist->getId(), $request->request->get('_token'))) {
             $entityManager->remove($artist);
             $entityManager->flush();
